Refactor fileevent package to separate FileAndName struct and add LogValue method
